+++ ContainerBuilder.php
@@ -1459,10 +1459,13 @@ class ContainerBuilder extends Container implements TaggedContainerInterface
      * using camel case: "foo.bar" or "foo_bar" creates an alias bound to
      * "$fooBar"-named arguments with $type as type-hint. Such arguments will
      * receive the service $id when autowiring is used.
+     *
+     * @param ?string $target
      */
-    public function registerAliasForArgument(string $id, string $type, ?string $name = null): Alias
+    public function registerAliasForArgument(string $id, string $type, ?string $name = null/*, ?string $target = null */): Alias
     {
         $parsedName = (new Target($name ??= $id))->getParsedName();
+        $target = (\func_num_args() > 3 ? func_get_arg(3) : null) ?? $name;
 
         if (!preg_match('/^[a-zA-Z_\x7f-\xff]/', $parsedName)) {
             if ($id !== $name) {
@@ -1472,8 +1475,8 @@ class ContainerBuilder extends Container implements TaggedContainerInterface
             throw new InvalidArgumentException(\sprintf('Invalid argument name "%s"'.$id.': the first character must be a letter.', $name));
         }
 
-        if ($parsedName !== $name) {
-            $this->setAlias('.'.$type.' $'.$name, $type.' $'.$parsedName);
+        if ($parsedName !== $target) {
+            $this->setAlias('.'.$type.' $'.$target, $type.' $'.$parsedName);
         }
 
         return $this->setAlias($type.' $'.$parsedName, $id);